// src/app/Model/quoteDetails.ts
export class QuoteDetails {
    checked:boolean;
    autorenew_id: string;
    bp_link_id: string;
    comments: string;
    contract_expiry_date: string;
    dealer_partner_name: string;
    expiring_contractnumber: string;
    expiring_ponumber: string;
    is_edited: string;
    is_eligible: string;
    is_promotion_applied: string;
    mea_id: string;
    owcc_id: string;
    parent_qrn: string;
    poStatus: string;
    poSubmitedBy: string;
    poSubmitedDate: string;
    po_amount: string;
    po_document: string;
    po_expiry_date: string;
    po_issue_date: string;
    po_number: string;
    quoteSubStatusName: string;
    quote_created_date: string;
    quote_end_date: string;
    quote_id: number;
    quote_start_date: string;
    quote_status: string;
    quote_status_name: string;
    quote_sub_status: string;
    reason: string;
    shipto: string;
    total_contract_value: string;
    avaya_net_price:string;
    t1name:string;
    t2name:string;
    res_owcc_id:string;
	res_po_amount:string;
	res_po_number:string;
	res_po_issue_date:string;
	res_po_submitted_by:string;
	res_po_submitted_date:string;
	res_po_expiry_date:string; 
 }